+++ SPMeta2/SPMeta2.CSOM/Services/CSOMIncrementalProvisionService.cs
@@ -24,7 +24,7 @@ namespace SPMeta2.CSOM.Services
 
         public CSOMIncrementalProvisionService()
         {
-            this.SetIncrementalMode();
+            this.SetIncrementalProvisionMode();
         }
 
         #endregion
